pyproject.toml
[tool.poetry]
name = "mipa-gamebot"
version = "0.1.0"
description = "AAA"
authors = ["tetsuya-k <64536338+tetsuya-ki@users.noreply.github.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
mipa = "^0.4.0"
python-dotenv = "^1.0.0"
requests = "^2.30.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
